+++ ChangeLog.md
@@ -149,3 +149,5 @@
 - Tests entfernt. Wurden nicht gepflegt, somit ist deren Nicht-Existenz die ehrlichere Lösung.
 
 ## Unreleased changes
+
+- Gtk-UI: Plan & Märklin-Weiche Hinzufügen in ScrollbaresWidget angezeigt.